refactor compare account
